+++ src/main/java/foro/hub/api/controller/TopicoController.java
@@ -1,8 +1,8 @@
 package foro.hub.api.controller;
 
 import foro.hub.api.domain.curso.CursoRepository;
-import foro.hub.api.domain.curso.DatosDetalleCurso;
 import foro.hub.api.domain.topico.*;
+import foro.hub.api.domain.topico.validaciones.ValidadorDeTopicos;
 import foro.hub.api.domain.usuario.UsuarioRepository;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
@@ -11,9 +11,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/topicos")
@@ -28,6 +31,9 @@ public class TopicoController {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
+    @Autowired
+    private List<ValidadorDeTopicos> validadores;
+
     @Transactional
     @PostMapping
     public ResponseEntity registroTopico(@RequestBody @Valid DatosRegistroTopico datos, UriComponentsBuilder uriComponentsBuilder) {
@@ -37,9 +43,12 @@ public class TopicoController {
         var curso = cursoRepository.findByNombre(datos.nombreCurso())
                 .orElseThrow(() -> new EntityNotFoundException("Curso no encontrado."));
         var topico = new Topico(datos, autor, curso);
+
+        validadores.forEach(v -> v.validar(datos));
+
         topicoRepository.save(topico);
 
-        var uri = uriComponentsBuilder.path("/cursos/i{id}").buildAndExpand(topico.getId()).toUri();
+        var uri = uriComponentsBuilder.path("/cursos/{id}").buildAndExpand(topico.getId()).toUri();
 
         return ResponseEntity.created(uri).body(new DatosDetalleTopico(topico));
     }
@@ -54,16 +63,40 @@ public class TopicoController {
 
     @Transactional
     @PutMapping("/{id}")
+    @PreAuthorize("""
+            hasAnyRole('MODERADOR', 'DOCENTE')
+            or (#topicoRepository.findById(#id).isPresent()
+            and #topicoRepository.findById(#id).get().getAutor().getId() == authentication.principal.id)
+            """)
     public ResponseEntity actualizarTopico(@PathVariable Long id, @RequestBody @Valid DatosActualizacionTopico datos){
-        var topico = topicoRepository.getReferenceById(id);
+        var topico = topicoRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Tópico no encontrado"));
         topico.actualizarInformacion(datos);
+
+        return ResponseEntity.ok(new DatosDetalleTopico(topico));
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity detallarTopico(@PathVariable Long id) {
+        var topico = topicoRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Tópico no encontrado"));
 
         return ResponseEntity.ok(new DatosDetalleTopico(topico));
     }
 
     @Transactional
     @DeleteMapping("/{id}")
-    public void eliminarTopico(@PathVariable Long id) {
-        topicoRepository.deleteById(id);
+    @PreAuthorize("""
+            hasAnyRole('MODERADOR', 'ADMIN')
+            or (#topicoRepository.findById(#id).isPresent()
+            and #topicoRepository.findById(#id).get().getAutor().getId() == authentication.principal.id)
+            """)
+    public ResponseEntity eliminarTopico(@PathVariable Long id) {
+        if (topicoRepository.existsById(id)) {
+            topicoRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            throw new EntityNotFoundException("Tópico no encontrado con el ID:" + id);
+        }
     }
 }
